// src/models/ResponseMovie.ts
import { GenresType } from "./MovieExtras";

export type ResponseMovie = {
  adult?: boolean;
  backdrop_path: string;
  genre_ids?: number[];
  id?: number;
  original_language: string;
  original_title: string;
  overview: string;
  popularity?: number;
  poster_path: string;
  release_date: string;
  title: string;
  video?: false;
  vote_average?: number;
  vote_count?: number;
  genres?: GenresType[];
  runtime?: number;
  tagline?: string;
  year?: string;
};